Revert "Feat/auth"
