+++ src/components/Categories/categories.js
@@ -1,11 +1,11 @@
-import React, { useContext } from 'react';
+import React, { useState, useEffect } from 'react';
 import { colors, typography } from '../../styles';
-import { handleAddTransaction } from '../../services/category-services';
-import CategoriesContext from '../../context/categoriesContext';
-import CategoriesHeader from '../CategoriesHeader/categories-header';
+import { getMonthlyData } from './utils';
+import apiFetch from '../../services/api-fetch';
 import CategoriesList from '../CategoriesList';
 import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
+import { AiOutlineMinusCircle, AiOutlinePlusCircle } from 'react-icons/ai';
 
 const Wrapper = styled.div`
   display: flex;
@@ -32,28 +32,72 @@ const TotalLabel = styled.p`
   font-weight: 500;
   color: ${colors.gray[500]};
 `;
+
+const StyledNavigator = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.6rem;
+  background-color: white;
+`;
+
+const StyledActions = styled.ul`
+  display: flex;
+  list-style: none;
+  padding: 0;
+  gap: 1rem;
+`;
+
+const StyledAction = styled.li`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
+  cursor: pointer;
+  color: ${colors.pink[400]};
+`;
+
+const Minus = styled(AiOutlineMinusCircle)`
+  font-size: 1.5rem;
+`;
+
+const Plus = styled(AiOutlinePlusCircle)`
+  font-size: 1.5rem;
+`;
 function Categories({ date, type }) {
-  const {
-    categories,
-    setCategories,
-    loading,
-    error,
-    filteredCategories,
-    setFilteredCategories,
-    monthlyData,
-  } = useContext(CategoriesContext);
+  const [categories, setCategories] = useState([]);
+  const [filteredCategories, setFilteredCategories] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  const monthlyData = getMonthlyData(categories, date, type);
 
   const total = monthlyData.reduce((acc, cur) => acc + cur.amount, 0);
 
-  function onAddTransaction(categoryId, data) {
+  function handleAddTransaction(categoryId, data) {
     const newCategories = [...categories];
     const category = newCategories.find((cat) => cat.id === categoryId);
 
-    handleAddTransaction(categories, categoryId, data).then((trx) => {
+    apiFetch(`categories/${category.id}/transactions`, {
+      body: data,
+    }).then((trx) => {
       category.transactions.push(trx);
       setCategories(newCategories);
     });
   }
+
+  useEffect(() => {
+    setLoading(true);
+    setError(null);
+    apiFetch('categories')
+      .then((data) => {
+        setCategories(data);
+        setLoading(false);
+      })
+      .catch((error) => {
+        setLoading(false);
+        setError(error);
+      });
+  }, []);
 
   if (loading) return <p>Loading categories...</p>;
   if (error) return <p>Error: {error.message}</p>;
@@ -74,11 +118,19 @@ function Categories({ date, type }) {
 
   return (
     <Wrapper>
-      <CategoriesHeader
-        title='Categories'
-        onPrev={handleShowExpenses}
-        onNext={handleShowIncomes}
-      />
+      <StyledNavigator className='category__navigator'>
+        <h1>Categories</h1>
+        <StyledActions className='category__navigator__actions'>
+          <StyledAction onClick={handleShowExpenses}>
+            <Minus />
+            Expenses
+          </StyledAction>
+          <StyledAction onClick={handleShowIncomes}>
+            <Plus />
+            Incomes
+          </StyledAction>
+        </StyledActions>
+      </StyledNavigator>
       <TotalWrapper>
         <TotalAmount>$ {Intl.NumberFormat('en-US').format(total)}</TotalAmount>
         <TotalLabel>
@@ -87,7 +139,7 @@ function Categories({ date, type }) {
       </TotalWrapper>
       <CategoriesList
         data={monthlyData}
-        onAddTransaction={onAddTransaction}
+        onAddTransaction={handleAddTransaction}
         onAddCatrgory={setCategories}
         categories={type === 'expense' ? categories : filteredCategories}
         date={date}
