--- src/components/CategoriesHeader/index.js
@@ -1,3 +1,0 @@
-import CategoriesHeader from './categories-header';
-
-export default CategoriesHeader;
