--- src/context/categoriesContext.js
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-const CategoriesContext = createContext({});
-
-export const CategoriesProvider = CategoriesContext.Provider;
-
-export default CategoriesContext;
