--- src/context/monthPickerContext.js
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-const MonthPickerContext = createContext({});
-
-export const MonthPickerProvider = MonthPickerContext.Provider;
-
-export default MonthPickerContext;
