--- src/context/navigationContext.js
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-const NavigationContext = createContext({});
-
-export const NavigationProvider = NavigationContext.Provider;
-
-export default NavigationContext;
