+++ src/pages/CategoryPage.js
@@ -1,13 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { useSearchParams, useParams } from 'react-router-dom';
 import { format, getMonth, getYear } from 'date-fns';
 import MonthPicker from '../components/MonthPicker';
-import { MonthPickerProvider } from '../context/monthPickerContext';
-import { CategoriesProvider } from '../context/categoriesContext';
 import styled from '@emotion/styled';
 import Categories from '../components/Categories/categories';
-import { handleGetCategories } from '../services/category-services';
-import { getMonthlyData } from '../components/Categories/utils';
 
 const StyledCategory = styled.div`
   margin: 24px 32px;
@@ -20,11 +16,6 @@ const StyledCategory = styled.div`
 `;
 
 function CategoryPage() {
-  const [categories, setCategories] = useState([]);
-  const [filteredCategories, setFilteredCategories] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
   let params = useParams();
 
   const [searchParams, setSearchParams] = useSearchParams({
@@ -53,46 +44,14 @@ function CategoryPage() {
     setSearchParams({ year, month });
   };
 
-  const monthlyData = getMonthlyData(categories, date, type);
-
-  useEffect(() => {
-    setLoading(true);
-    setError(null);
-    handleGetCategories()
-      .then((data) => {
-        setCategories(data);
-        setLoading(false);
-      })
-      .catch((error) => {
-        setLoading(false);
-        setError(error);
-      });
-  }, []);
-
   return (
     <StyledCategory className='category'>
-      <MonthPickerProvider
-        value={{
-          label: format(new Date(date.year, date.month), 'MMMM yyyy'),
-          onLeftClick: handleLeftClick,
-          onRightClick: handleRightClick,
-        }}
-      >
-        <CategoriesProvider
-          value={{
-            categories,
-            setCategories: setCategories,
-            filteredCategories,
-            setFilteredCategories: setFilteredCategories,
-            loading: loading,
-            error: error,
-            monthlyData,
-          }}
-        >
-          <MonthPicker />
-          <Categories {...{ date, type }} />
-        </CategoriesProvider>
-      </MonthPickerProvider>
+      <MonthPicker
+        label={format(new Date(date.year, date.month), 'MMMM yyyy')}
+        onRightClick={handleRightClick}
+        onLeftClick={handleLeftClick}
+      />
+      <Categories {...{ date, type }} />
     </StyledCategory>
   );
 }
