+++ src/services/api-fetch.js
@@ -1,10 +1,11 @@
-import { BASE_URI, tokenKey } from '../config';
+import { BASE_URI } from '../config';
 
 export default async function apiFetch(
   endpoint,
   { method, headers, body } = {}
 ) {
-  const token = sessionStorage.getItem(tokenKey);
+  // const token = sessionStorage.getItem(tokenKey);
+  const token = 'KpNeEJQfQRBzDSiwz4x2fAyD';
 
   if (token) {
     headers = {
