--- src/services/session-services.js
@@ -1,15 +1,0 @@
-import { tokenKey } from '../config';
-import apiFetch from './api-fetch';
-
-export async function login(credentials) {
-  const { token, ...user } = await apiFetch('login', {
-    body: credentials,
-  });
-  sessionStorage.setItem(tokenKey, token);
-  return user;
-}
-
-export async function logout() {
-  await apiFetch('logout', { method: 'DELETE' });
-  sessionStorage.removeItem(tokenKey);
-}
